1st step of the project implementation and tests (#2)
The 1st step mainly consists of going from

`new StringContext("", "%d").format(x) `
to
`"%d".format(x)`

This pull request adds the implementation of this and the tests.

Remark : some floats are not stored as we think they are. Namely, 1234,5678 is stored as 1234,56774 which makes the formatter output something we do not expect. This is a simple loss of precision due to numerical storing methods.